requirements.txt
Django==1.9.5
djangorestframework==3.3.3
Markdown==2.6.6
django-filter==0.13.0
django-crispy-forms==1.6.0
django-guardian==1.4.4
djangorestframework-jwt==1.8.0
django-cors-headers==1.1.0
qrcode==5.2.2